% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apobec_enrichment.R
\name{apobec_enrichment}
\alias{apobec_enrichment}
\title{Calculate APOBEC enrichment}
\usage{
apobec_enrichment(inputdata)
}
\arguments{
\item{inputdata}{a dataframe with columns of twentycontext (20 nucleotide context),
con (trinucleotide context), sample (Sample names), mut_TCW (TCW context mutations),
MutC (mutations at C) This dataframe is created using either read_nonvcf function or
vcfs_to_inputdata function}
}
\value{
enrichment table with adjusted p-values
}
\description{
Calculate APOBEC enrichment
}
