% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_signatures.R
\name{assign_signature}
\alias{assign_signature}
\title{Find optimal nonnegative linear combination of mutation signatures to
reconstruct the mutation matrix using nnls package.}
\usage{
assign_signature(known_sig_mat, mat, cut_off = 0.03)
}
\arguments{
\item{known_sig_mat}{Known Signature matrix (dimensions: 96 mutation types
30 n signatures)}

\item{mat}{mutation count matrix (dimensions: 96 mutation types
X n samples)}

\item{cut_off}{Minimum signature contribution required}
}
\value{
Named list with signature contributions and reconstructed
mutation matrix
}
\description{
Find optimal nonnegative linear combination of mutation signatures to
reconstruct the mutation matrix using nnls package.
}
\examples{


## Get signatures
signatures <- get_known_signatures()

## Perform the fitting
fit <- assign_signatures(known_sig_mat, mat, cut_off=0.03)

}
