% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRC.R
\name{exe_fisher}
\alias{exe_fisher}
\title{Execute one-sided Fisher's exact test}
\usage{
exe_fisher(x)
}
\arguments{
\item{x}{data frame with elements of contingency table as columns and samples as rows}
}
\value{
Fisher's exact p-calue
}
\description{
Execute one-sided Fisher's exact test
}
