% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRC.R
\name{getRC}
\alias{getRC}
\title{Get the reverse complement of a DNA string}
\usage{
getRC(s)
}
\arguments{
\item{s}{DNA sequence to be converted}
}
\value{
reverse complement sequence string
}
\description{
et the reverse complement of a DNA string
}
