% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutmat_nonvcf.R
\name{mutmat_nonvcf}
\alias{mutmat_nonvcf}
\title{Make mutation count matrix of 96 trinucleotides from a dataframe with contexts and samplenames
for each mutation (this data frame is created using read_nonvcf)}
\usage{
mutmat_nonvcf(inputdata)
}
\arguments{
\item{inputdata}{Dataframe containing con (context column) and sample column to form the mutation matrix}
}
\value{
96 mutation count matrix
}
\description{
Make 96 trinucleotide mutation count matrix
}
