% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sigs.R
\name{plot_sigs}
\alias{plot_sigs}
\title{Plot signature contribution barplot}
\usage{
plot_sigs(fit, mode = c("relative", "absolute"), signatures = NA)
}
\arguments{
\item{fit}{Signature contribution matrix (Fit$coeffecient from assign_signatures)}

\item{mode}{"relative" or "absolute"; to plot the relative contribution or
absolute number of mutations, default = "relative"}

\item{signatures}{Signature matrix. Default is NA}
}
\value{
Stacked barplot with contribution of each signature for each sample
}
\description{
Plot contribution of known signatures. Adopted from plot_contribution
from MutationalPatterns
}
\examples{


## Plot the relative contribution
plot_sigs(fit$coefficient, mode="relative")

## Plot the absolute contribution.
## When plotting absolute NMF results, the signatures need to be included.
plot_sigs(fit$coefficient,
  mode = "absolute"
)


}
