% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nonvcf.R
\name{read_nonvcf}
\alias{read_nonvcf}
\title{Read a non-vcf mutation file to create the inputdata frame.}
\usage{
read_nonvcf(inputdata, bsg = bsg, ext = 1)
}
\arguments{
\item{inputdata}{a dataframe with columns chr, start, stop, ref, alt, sample}

\item{bsg}{reference BSGenome object to map}

\item{ext}{extention of context (current only 1 is supported)}
}
\value{
dataframe with columns displaying 20 nucleotide context, 3 nucleotide context
}
\description{
Read a non-vcf mutation file to create the inputdata frame.
}
