% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfs_to_inputdata.R
\name{vcfs_to_inputdata}
\alias{vcfs_to_inputdata}
\title{Read a vcf mutation file to create the inputdata frame.
Uses read_vcf function from sigminer package}
\usage{
vcfs_to_inputdata(
  vcffiles,
  bsg = bsg,
  ext = 1,
  samples = sample_names,
  genome = gen
)
}
\arguments{
\item{vcffiles}{a vcf file list}

\item{bsg}{bs genome object}

\item{ext}{extention of context (current only 1 is supported)}

\item{samples}{sample name list}

\item{genome}{one of "mm10" or "hg38" description}
}
\value{
dataframe with columns displaying 20 nucleotide context, 3 nucleotide context
}
\description{
Read a vcf mutation file to create the inputdata frame.
Uses read_vcf function from sigminer package
}
